fix: bug use Params , Hook Suspense, export default
